requirements.txt
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
numpy==1.26.4
torchvision==0.21.0+cu124
transformers==4.52.3
huggingface-hub==0.36.0
peft==0.11.1
accelerate==1.6.0
datasets==3.6.0
scipy==1.15.3
tqdm==4.67.1
wandb==0.17.7
pillow==10.4.0
wrapt==1.17.2
py-cpuinfo==9.0.0
hjson==3.1.0
scikit-learn==1.7.0
scikit-image==0.25.2
ray==2.47.1
timm==1.0.15
sentencepiece==0.2.0
opencv-python==4.11.0.86
decord==0.6.0
hnswlib==0.8.0
einops==0.8.1
ninja==1.11.1.4
opencv-contrib-python==4.11.0.86
flash_attn==2.6.3  # check out orginal repo README for accelerating installation: https://github.com/Dao-AILab/flash-attention
qwen-vl-utils[decord]==0.0.8  # https://huggingface.co/Qwen/Qwen2.5-VL-3B-Instruct/discussions/30